+++ Solver/Exercise.h
@@ -7,8 +7,8 @@ namespace miit::algebra
 	class Exercise :public miit::algebra::VirtualExercise
 	{
 	public:
-		Exercise(int rows, int columns, Generator& generator);
+		Exercise(size_t rows, size_t columns, Generator& generator);
 		void task1() override;
 		void task2() override;
-	};
+			};
 }
